backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
google-generativeai==0.3.2
PyJWT==2.8.0
bcrypt==4.1.2
aiosqlite==0.19.0
httpx==0.26.0
nltk==3.8.1
scikit-learn==1.4.0
numpy==1.26.3
